Add the ability to start and stop pools
This commit replaces the previous approach of unlocking encrypted pools
with full support for starting and stopping pools. Locked encrypted
pools are now simply considered stopped.

The behavior of stopping pools is as follows. A pool that is started
when the system shuts down will be started after booting the system
again. A pool that is stopped will remain stopped and will not be set up
until it is started again. This operation requires a change to the
metadata schema to store whether a pool is stopped or started. Migration
from older versions of stratisd is handled.

Stopping an encrypted pool will lock all of the encrypted devices.