Fix potential deadlock in AllOrSomeLock
This potential deadlock occurs when a pool cannot be found and there are
other tasks waiting to acquire the lock. The read and write methods
return None. Because there is no guard acquired to later be dropped to
wake up the waiting tasks, the waiting tasks are never woken up and
every subsequent request gets queued forever.

This change adds a wake up when the pool is not found because this
indicates no lock could be acquired and thus any other waiting tasks are
now free to proceed with acquisition.