Increase the initial size of the MDV to 512 MiB
Increase the pool size for some tests so that the test doesn't run out
of space while creating filesystems.

Do not pass the noalign flag when creating the MDV any more. It should
be big enough that it is no longer necessary to pass the option to
ensure that the MDV is created.

Signed-off-by: mulhern <amulhern@redhat.com>