Remove test macro that checks for expected filesystem size
It is obsoleted now that it is possible to create filesystems
of many sizes.

At bottom, it is probably flakey. It has been very consistent in our
CI test infrastructure, but we can not guarantee that layout effects would
not become apparent on different devices. So it should be omitted from
tests that we run in our CI, anyway.

Signed-off-by: mulhern <amulhern@redhat.com>