Remove dbus_api condvar
Previously, this condvar allowed the D-Bus API to wait on pool creation.
This introduced some problems that can be better handled by using the
AllOrSomeLock internally in the engine to hold an exclusive lock on all
pools until the pool is created and inserted into internal data
structures. This:
* prevents duplicate pools as liminal devices won't be able to process
the udev events from pool creation until the pool is already set up
when they will then be ignored
* removes the need for any synchronization between the engine and the
D-Bus layer
* prevents the Drop trait from being run on the pool every time a pool
is created because we previously immediately dropped it after setting it
up (this behavior would cause problems if the MDV is kept mounted for
the entire lifetime of the pool)

This change dramatically simplifies pool creation to be closer to our
original state prior to adding async support at the engine level and
gives us strong ordering guarantees like the pool being in internal data
structures by the time create_pool completes.