Drop udev_settle() calls in filesystem and thinpool
Remove the two invocations of udev_settle() in the clean up paths of
Stratfilesystem::initialize() and Thinpool::create_filesystem() and
replace them with a limited retry using the retry crate's
retry_with_index().

Since the function is now unused in both files (outside test code)
remove the use statements for the function as well: this leaves the
function without any callers in production code, although there are
numerous test cases that use it: mark the function definition with
the #cfg(test) annotation.

Set the maximum retries for both cases to 4 (i.e. 5 total attempts).

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>