Fix FailDevice.start_failing() mapping offset
The FailDevice.start_failing() method accepts a num_sectors_after_start
argument that controls how many sectors from the start of the device
should be replaced with the error mapping. This is used to construct a
dm table with two segments (error and linear) that replaces the original
one segment linear mapping:

Original:

  stratis_fail_device: 0 2097152 linear 252:19 0

After start_failing():

  stratis_fail_device: 0 32768 error
  stratis_fail_device: 32768 2064384 linear 252:19 0

Although it does not affect the primary purpose of the error mapping (to
trigger read errors for IO to the LUKS header of the device), this
shifts the content of the mapped device +16M: sector 0 of the original
mapping becomes sector 32768 of the failing device. If this part of the
device address space contains file system or thin pool content it will
appear to be corrupted (and writes may corrupt the underlying content).

This happens in the Stratis test suite when testing with small device
sizes (~1GiB) and the clevis_real_test_clevis_rebind_rollback and
clevis_loop_test_clevis_rebind_rollback test cases. The MDV for the test
pool can end up spanning the beginning of the fail device resulting in a
shifting of part of the fs address space and corruption of the journal
that results in EINVAL when attempting to mount the volume:

[2022-07-06T16:51:33Z INFO  stratisd::engine::strat_engine::liminal::liminal] Attempt to set up pool failed, but it may be possible to set up the pool later, if the situation changes: An attempt to set up pool with UUID 885e831f-55b5-4f4f-b26a-8624882c07f8 from the assembled devices failed; Nix error: EINVAL: Invalid argument
...
thread 'engine::strat_engine::engine::test::clevis_real_test_clevis_rebind_rollback' panicked at 'called `Result::unwrap()` on an `Err` value: NoActionRollbackError { causal_error: Chained("An attempt to set up pool with UUID 885e831f-55b5-4f4f-b26a-8624882c07f8 from the assembled devices failed", Nix(EINVAL)), rollback_error: Chained("Failed to roll back encrypted pool unlock; some previously locked encrypted devices may be left in an unlocked state", Crypt(IOError(Os { code: 16, kind: ResourceBusy, message: "Device or resource busy" }))) }', src/engine/strat_engine/engine.rs:1132:10

Fix this by also applying the +16M offset to the linear mapping so that
the correct content appears to users of the mapped device:

  stratis_fail_device: 0 32768 error
  stratis_fail_device: 32768 2064384 linear 252:19 32768

This resolves the failure seen with the two Clevis rollback test cases
and small test device sizes.