Fix bug in mutability guarantees for locking data structure
While this bug is never exploited in our code and the data structure
cannot be used outside of this crate, SomeLockWriteGuard did not uphold
the mutability guarantee that Rust provides. The use of an immutable
reference in the .as_tuple() could be called multiple times, generating
multiple valid mutable references to pool with no way for the compiler
to prevent this.

Changing the method to .as_mut_tuple() that takes &mut self now ties the
mutable reference returned from the method to the &mut self reference,
enabling the compiler to once again verify that only one mutable
reference can be taken on the pool at a time.