Improve header write API
Before this commit, the closure in repair_sigblock would return
StratisResult<Option<StaticHeader>>. While this provides increased
flexibility for closures to determine a device as "not a Stratis
device", the code preceding the invocation of the closure actually
already determines whether there is at least one valid header or not on
the device. As a result, it doesn't make much sense to decide after that
check has already been performed.

The motivation for this change is that if writing the header is
performed outside of the repair_sigblock method, there is an unreachable
condition of the return type being None which has to be handled. This
generates an unnecessary expect when this could simply be eliminated by
removing the optional return type.