Increase chrono dependency lower bound to 0.4.20
This way we avoid the message from "cargo audit". We had been able to
prove that stratisd was not affected by the vulnerability, see
https://github.com/stratis-storage/project/issues/349#issuecomment-963132879,
but since an actual fix was put in the warning has been re-enabled. The
crate is available in Fedora, so there is no problem in bumping the
lower bound.

Signed-off-by: mulhern <amulhern@redhat.com>