Increase pretty-hex dependency lower-bound to 0.3.0
Ignore that the requirement is too high for Fedora 35. We get that
dependency from stratis-dumpmetadata, which we do not yet release and
do not expect to release in Fedora 35.

Note that it would be better if the dependencies we could ignore could
be selected automatically by using the features flag to cargo-metadata
to specify only the features that are required for the various binaries
that we actuall release. Unfortunately, even when --features args are
used to specify a particular subset of features, cargo-metadata output
still interprets pretty-hex as a dependency. This is another argument
for switching to Rust. Using the metadata data structures available in
the Rust crate should make it easy to automatically eliminate
pretty-hex from consideration as a dependency for Fedora release
purposes.

Signed-off-by: mulhern <amulhern@redhat.com>