Restrict bindgen features to remove clap dependency
Note that for some reason, cargo audit still reports the CVE on
stratisd.

Note also that setting the edition to 2021 does not change this
behavior. I had though it would be changed, since edition 2021 resolves
dependencies with different features separately.

But it seems that cargo audit interprets all the bindgen dependencies as
identical, even though devicemapper's bindgen dependency has its
features restricted so that it has no clap dependency, evenn when the
edition is set to 2021.

Signed-off-by: mulhern <amulhern@redhat.com>