Fix for assertion when migrating from greedy to lazy allocation
The calculation for metadata size when migrating a pool from a version
of stratisd with greedy allocation to a version with lazy allocation did
not appropriately handle the case where the metdata device is already
larger than it needs to be in the new scheme.

This commit changes the metadata size calculation to handle the
outlined case with metadata size being larger than is needed.