Remove questionable TODO
It's not clear to me that failing pool setup if filesystem data contains
duplicate filesystems is the wrong thing to do.

It's also a problem we have yet to encounter, as far as I can recall.

We can refine our approach if we encounter the problem in the real
world.

Signed-off-by: mulhern <amulhern@redhat.com>