Replace SetUserInfo with settable property and log on property set
This commit addresses three separate issues.

One is that of SetUserInfo. This was inconsistent with our new approach
of settable properties. This commit replaces SetUserInfo with a settable
property for setting the user information on a given block device
exposed on the D-Bus.

Another is that setting properties did not log when the values were
changed. This is valuable logging information that is now included.

The last issue that this resolves is that setting the filesystem limit
previously was not idempotent. This is now resolved and returning an
error message should be the job of the CLI when the value does not
change.