Remove two calls to ThinPool::check
Each call is issued from a StratPool method, setup or add_blockdevs.

In both cases, each call was added before multi-threading was added to
stratisd. In each case, the call was used to react to a change that
would not cause a DM or udev event that stratisd could react to.

Now that a separate timer thread, which invokes the ThinPool::check
method every 10 seconds exists, the original motivation for invoking
ThinPool::check in these two methods has been removed.

In the case of add_blockdevs() it should not be necessary to guarantee
that the thinpool is expanded before the add_blockdevs method returns.
Some reasonably short time after that ought to be entirely sufficient.
In fact, omitting the thinpool check is a cleaner design, as it restricts the
consequences of adding blockdevs to changes in the backstore, which would be
more in line with what a user might be expecting to detect.

In the case of StratPool::setup(), the pool is being set up from its
prior state. If the prior state was not actually out of data- or meta-
space then the current state will not be either. It seems
extraordinarily unlikely that a thinpool check will "rescue" a
previously set up pool from an out of space condition that it was
previously in. We have observed the thinpool check in this context to do
a meta-device expansion, but that was in the case where the
stratisd version and metadata allocation algorithm changed. A change
such as that should be able to wait until the timer thread kicks off or
an event is detected, whichever comes first.

Signed-off-by: mulhern <amulhern@redhat.com>