Add optional encryption-related parameters
The optional CryptParamLuks2 parameter added to the initialize_with_err()
method will allow specifying the sector size with which to encrypt when
initializing a crypt device.

The sector_size parameter, if present, requires that an optional
CryptoParamsLuks2 object be supplied to the initialize_with_err()
method.

Push the optional sector_size parameter up to
devices::initialize_devices.

Signed-off-by: mulhern <amulhern@redhat.com>