Record sector sizes of crypt devices
Add a StratBlockSizes struct. Obtain crypt block sizes in
StratBlockDev::new. This is invoked both from liminal devices on setup and
when initializing a new device.

Modify invalidate() so that it returns a structure that includes
information about sector sizes of crypt devices.

Calculate a sector_size argument for initialization of block devices
when devices are added or a cache is initialized.

Signed-off-by: mulhern <amulhern@redhat.com>