On UnlockPool, start the pool directly
Relying on D-Bus events will not work anymore because an attempt to set
the pool up is made each time a udev event is encountered, and when it
fails, any opened crypt devices are torn down.

*Make StratEngine::unlock_pool like StratEngine::start_pool(), but strip
out the irrelevant parts.
*Require liminal_devices::start_pool to return unlocked uuids, along with
other results.
* Make SetUnlockAction an enum so that it is possible to tell if a pool
has been started as a result of the unlock action or not. There are two
cases where no devices are unlocked: 1) when there was a failure to tear
down all devices previously (for some reason) but an attempt to start
the pool succeeds this time, 2) where the pool is an encrypted pool that
has already been started. (2) corresponds to the Identity constructor,
(1) to the Started constructor with an empty list.

Signed-off-by: mulhern <amulhern@redhat.com>