Increase rpassword dependency lower bound to 7.2.0
Substitute prompt_password for removed prompt_password.
Note that prompt_password prints the prompt string to /dev/tty, while
prompt_password_stdout printed the prompt to stdout.

Remove test that prints password prompt. Since the prompt goes to tty
rather than stdout, the test framework is not able to handle it. This is
a know limitation of the framework, see:
https://github.com/assert-rs/assert_cmd/issues/138

Signed-off-by: mulhern <amulhern@redhat.com>