Ensure persistent keyring is attached before Clevis bind
Recently, the memory mapped keyfile workaround for Clevis was removed.
In removing this code, we left a code path open where binding using
Clevis could result in an error if stratisd is newly started and stratis
key list is not run prior to running the bind command. This change
ensures that even if stratisd is newly started and then a clevis bind
command is run, the key will be found in the keyring.