Remove dm::stop_partially_constructed_pool
In stratisd 3.6.0 stop_partially_constructed_pool() is used via a D-Bus
StopPool call which allows the user to explicitly request that a
partially set up pool be stopped. In stratisd < 3.6.0 this can not be
done, so, when the teardown of a partially constructed pool on setup is
eliminated in liminal.rs, the method becomes dead.

This change broke the D-Bus API, so we can not add it in in this branch.

Signed-off-by: mulhern <amulhern@redhat.com>