Revert allocation transaction infrastructure
This code has two notable drawbacks that make it better to remove.

1. The code is never used as intended. All requests are immediately
followed by a commit. This does not justify the complexity required to
maintain the feature.
2. There is an assumption that will likely be broken in future commits
related to changes in stacking and metadata. The assumption is that
allocations are always driven by the backstore. If this assumption is
broken and allocations to reserve space for metadata are performed at
layers lower than the backstore, these layers don't have protections
against reserving the same space for two separate allocations and then
commiting them both.