Relax fairness restriciton on AllOrSomeLock
This commit relaxes the previous problematic restriction that was meant
to enforce fairness at the cost of throughput. Previously all waiting
lock acquisitions were queued and if there was a conflicting lock
acquisition before a lock acquisition that could currently be served, it
would still be left queued. This commit relaxes that restriction due to
its high potential to cause deadlocks in stratisd.