Fix PID file bug in stratisd and stratisd-min
This commit fixes a bug in the Stratis daemons. The bug would
theoretically only occur if stratisd gets restarted and reassigned a PID
with less digits than the previous PID. This commit truncates the file
when we determine there is no other instance of the daemon running to
avoid this occuring.