Specify PathBuf type for blockdevs in parser
Note that this does virtually nothing. A PathBuf is basically just an
OsString with extra behaviors. And it can always be parsed from a
String, which is the clap default argument value type.

All this change accomplishes is that when processing the argument values,
the paths will not have to be converted to Strings and then Pathbufs
because that conversion happened when they were read, because clap was
instructed to do it.

This change just lays the foundation for good practice where it might be
noticeable, so that when clap makes it possible, other values, e.g., UUIDs,
can be checked for proper format by the parser, and if the format is wrong,
a parser error will be returned as part of the parser operation.

Signed-off-by: mulhern <amulhern@redhat.com>
(cherry picked from commit 5648d20add209107cd5bc1d167fc94355180daa9)