Move shared.rs to dispatch.rs
The naming helps make it clear what it is for. Where we use shared.rs
elsewhere, we use sharing in the opposite sense, so a different word
helps here.

Also, we may at sometime find a good way to isolate some bits of the
shared pool implementation and it would cause confusion if they didn't
go in a shared.rs file.

Signed-off-by: mulhern <amulhern@redhat.com>