Read the filesystem metadata before any setup
Verify that scheduled merges are permissible.

Return with an error without setting up any filesystems if duplicate
UUIDs or names are found.

Perform the merges before setting up any filesystems. If a merge can be
rolled back, set up the two filesystems in the merge relation in the
normal way.

Signed-off-by: mulhern <amulhern@redhat.com>