For idempotency criterion, check only equality of pins
Do not strip config of thumbprint entry, as the comparison that required
that operation has been removed.

It would be nice to also check equality of URLs, if the pin is "tang".
But that is a lot of work for a small benefit. If a user for some reason
attempts to create a pool with the same name, the same devices, and the
same pin, as the pool was created with previously, then
stratisd will return a value indicating that the pool was already
created even if something else about the Clevis config is different.
This seems acceptable.

Signed-off-by: mulhern <amulhern@redhat.com>