No longer use interpret_clevis_config in backstore
The return value was used to decide whether to remove a key/value pair,
but that no longer needs to be removed.

The argument passed to it was only used for a comparison which is no
longer being made and an error message display that can use the
value it was cloned from instead.

If the config was malformed, then it would error early, i.e., before
attempting to bind the first blockdev. Now it will just error when it
tries to bind the first blockdev. This is simpler to reason about.

Signed-off-by: mulhern <amulhern@redhat.com>