Make D-Bus signature for integrity tag spec a string
Define an enum to represent the tag size.

Store the IntegrityTagSpec value in the pool-level metadata.
Nothing is gained by turning the enum into bits before storing it in the
pool-level metadata.

Compute using the bytes value that is at least as large as the number of
bits the tag spec represents.

Use tag_spec instead of tag_size for the name of the field in the
CreatePool method.

Signed-off-by: mulhern <amulhern@redhat.com>