Fix how keyring keys are provided to Clevis
Previously keyring keys would be passed to Clevis via stdin. This is a
problem, because keyring keys should theoretically be able to have
newline characters if passed in via a keyfile. stdin would truncate the
passphrase at the first newline on the Clevis side, resulting in an error
binding to Clevis for any passphrase that has a newline before the end of
the passphrase.