Use clap cargo feature to print authors
Requires the use of a help_template to include authors in help text, so
this is not obviously worth it. The current template only shows the
authors, and is not a serious suggestion.

Other macros are crate_name and crate_description which are not
generally useful for us since we build so many executable and
crate_version, which is perhaps useful.

Signed-off-by: mulhern <amulhern@redhat.com>