Trim password before checking for emptiness
read_line() will include the newline in the value read if that is what
terminates the string.

Previously, any newline-terminated string was interpreted as non-empty,
and its trimmed value, which could be the empty string, was
considered a valid password.

If the password that would be sent is empty, and the purpose is to set a
key in the kernel keyring, return an error without communicating over
the IPC.

However, if the purpose is to start the pool, send the empty passphrase
anyway. In the context where starting a pool is partially controlled by
systemd, exiting without transmitting the IPC instruction to stratisd
results in systemd not detecting any action, and thus repeating the
request for a passphrase endlessly, even if the nofail option is set.

Signed-off-by: mulhern <amulhern@redhat.com>