Trim password before checking for emptiness
read_line() will include the newline in the value read if that is what
terminates the string.

Previously, any newline-terminated string was interpreted as non-empty,
and its trimmed value, which could be the empty string, was
considered a valid password.

If the password that would be sent is empty, return an error without
communicating over the IPC.

Trim only terminating newline from user-entered passphrase
Leave all other whitespace as is. This means that a user-entered
passphrase consisting only of space is a legal passphrase.

Signed-off-by: mulhern <amulhern@redhat.com>