Reduce crypt log level error to info
For crypt it is an error, but for stratisd it is only a temporary
setback.

Best to avoid putting an error log entry in stratisd log messages and causing
undue alarm.

Changes:
[2025-03-05T22:49:56Z ERROR  stratisd::engine::strat_engine::crypt::shared] Failed to retrieve Clevis passphrase; Error communicating with server http://127.0.0.1

to

[2025-03-05T22:49:56Z INFO  stratisd::engine::strat_engine::crypt::shared] Failed to retrieve Clevis passphrase; Error communicating with server http://127.0.0.1

For the change to the logging_callback method I do not have an example
of a log entry set at too high a level, but I think that the same argument
still holds.

Signed-off-by: mulhern <amulhern@redhat.com>