Fix Makefile so that auditable checks can be run
The if section to run the various cargo components with minimal-versions if
MINIMAL environmental variable was set erased the effect of the previous
check that ran some cargo components with the audtitable tool if the
AUDITABLE environment variable was set.

This problem is fixed through the use of nested ifs which continue to
ensure that if neither environment variable is set the cargo components
are run in the normal way.

Signed-off-by: mulhern <amulhern@redhat.com>