Makefile: allow unmaintained advisory for humantime
We can not remove our dependency on humantime at this time.

We have this dependency through env_logger.

If we dependend on env_logger without default-features we still depend
on humantime because we also depend on env_logger through devicemapper.
In any case, dropping the humantime feature in env_logger seems quite
wrong, since it is used to make timestamps human readable.

We can not, at present, update our dependency to the newest version of
env_logger, which has dropped its dependency on humantime. cargo update
will not bump the version in Cargo.lock, presumably for other dependency
reasons.

In any case, humantime becoming officially unmaintained is no emergency; it has
probably been actually unmaintained for quite some time.

We expect to have dropped our dependency on humantime in the next
release.

Signed-off-by: mulhern <amulhern@redhat.com>