Unconditionally save pool level metadata after starting a pool
There are an accumulating set of reasons to save the metadata, some
complex. There is a danger that one will be overlooked.

Pools are rarely set up, so unconditionally writing the metadata when
setting up a pool will not have an overall performance impact in normal
use.

This PR is motivated by the consideration that when adding functionality
to set up a pool without its cache, it will likely be necessary to add
an additional parameter to the setup() method to control whether it
needs to write the metadata because it is skipping the cache. That seems
like one too many parameters, and so now is a good time to switch to
unconditional metadata writes.

In any case, whenever a pool is started from a stopped state, its
pool-level metadata must be written anyway.

Signed-off-by: mulhern <amulhern@redhat.com>