In error message, list the snapshots that must be changed
Previously, if it was discovered that a filesystem to be destroyed was
scheduled to be reverted to one of its snapshots, it was the UUID of the
filesystem to be destroyed that was included in the error message. It makes
more sense to include the UUIDs of the snapshots that are scheduled to be
merged, since the user must revert the scheduled changes to those
snapshots in order to be able to destroy the filesystem.

Given the UUID of any snapshot, the user can find out its origin by
listing the snapshot in the detail view. But given the UUID of an origin,
finding its snapshots is a reverse lookup, and more time consuming.

Signed-off-by: mulhern <amulhern@redhat.com>