Fix assertion error when calculating error message
Formerly, if a device with Stratis metadata was added to a pool, and the
pool's UUID matched the device's UUID, but the pool had no record of the
device, calculating the error to return resulted in a failed assertion
and a panicked thread. The visible result to the user was that their
request timed out, with no response.

thread 'stratis-wt-51' panicked at src/engine/strat_engine/pool/v2.rs:572:51:
was looked up

Signed-off-by: mulhern <amulhern@redhat.com>
(cherry picked from commit 84ebb5687df40df61d58d252bbc999e475ae0d95)