Remove all weekly GitHub workflows
These check whether the version of Python used in the next version of
Fedora may cause problems with our Python tools; but we check that in
other ways now.

Keeps a log of all the audit failures on previous versions of stratisd.
This is kind of interesting, but we have never taken any action based on
this data and likely never will.

stratisd always passes the audit check when it is released.

Signed-off-by: mulhern <amulhern@redhat.com>