No longer exclude vendor subdirectory in libdbus-sys
The problematical files in that directory are no longer included in the
crate.

We do not use the vendored feature, because we do not turn it on for the
dbus crate.

However, if we drop a configuration instruction to cargo-vendor-filterer
we are that bit less fussy and dependent on it.

Signed-off-by: mulhern <amulhern@redhat.com>