Duplicate file descriptor in set_key_method
Otherwise it gets closed twice, once by stratisd because it makes a File
struct from the raw pointer, and once by the zbus crate.

When it is closed twice in a release build there is generally no harm
done; but when in a debug build, there is an immediate abort.

See https://github.com/stratis-storage/stratisd/issues/3946 for further
details.

Signed-off-by: mulhern <amulhern@redhat.com>